+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Changes
+
+-   Added `isTenantIdDefined` to multitenancy
+
 ## [0.7.1] - 2023-07-24
 
 ### Changes